Adds uses to all Wagtail models.

+++ wagtail/images/views/images.py
@@ -10,6 +10,7 @@ from django.views.decorators.vary import vary_on_headers
 from wagtail.admin import messages
 from wagtail.admin.forms import SearchForm
 from wagtail.admin.utils import PermissionPolicyChecker, permission_denied, popular_tags_for_model
+from wagtail.core.collectors import get_paginated_uses
 from wagtail.core.models import Collection, Site
 from wagtail.images import get_image_model
 from wagtail.images.exceptions import InvalidFilterSpecError
@@ -226,13 +227,16 @@ def delete(request, image_id):
     if not permission_policy.user_has_permission_for_instance(request.user, 'delete', image):
         return permission_denied(request)
 
-    if request.method == 'POST':
+    uses = get_paginated_uses(request, image)
+
+    if request.method == 'POST' and not uses.are_protected:
         image.delete()
         messages.success(request, _("Image '{0}' deleted.").format(image.title))
         return redirect('wagtailimages:index')
 
     return render(request, "wagtailimages/images/confirm_delete.html", {
         'image': image,
+        'uses': uses,
     })
 
 
@@ -264,15 +268,4 @@ def add(request):
 
     return render(request, "wagtailimages/images/add.html", {
         'form': form,
-    })
-
-
-def usage(request, image_id):
-    image = get_object_or_404(get_image_model(), id=image_id)
-
-    paginator, used_by = paginate(request, image.get_usage())
-
-    return render(request, "wagtailimages/images/usage.html", {
-        'image': image,
-        'used_by': used_by
     })
